requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.0
PyYAML>=6.0.0
tqdm>=4.66.0

